Plant Disease  Detection/requirements.txt
Flask==2.3.2
matplotlib==3.5.2
numpy==1.23.1
pandas==1.5.0
Pillow==9.2.0
torch==2.0.1
torchsummary==1.5.1
torchvision==0.15.2
